refactor: consolidate ipg tests
